Reorganize namespace
Create a slightly more hierarchical namespace:

- ltdconveyor.fastly
- ltdconveyor.s3

This should give more space to grow ltd-conveyor's functionality as a
general LTD client library in the future.
