+++ docs/index.rst
@@ -38,5 +38,11 @@ Travis CI creates the PyPI release itself and `versioneer <https://github.com/wa
 API reference
 =============
 
+.. automodapi:: ltdconveyor.fastly
+   :no-inheritance-diagram:
+
+.. automodapi:: ltdconveyor.s3
+   :no-inheritance-diagram:
+
 .. automodapi:: ltdconveyor
    :no-inheritance-diagram:
