+++ ltdconveyor/s3/__init__.py
@@ -1,0 +1,6 @@
+from .copy import copy_dir
+from .delete import delete_dir
+from .exceptions import S3Error
+from .upload import (upload_dir, upload_file, upload_object,
+                     create_dir_redirect_object, ObjectManager)
+from .utils import open_bucket
