+++ ltdconveyor/s3/exceptions.py
@@ -1,0 +1,8 @@
+__all__ = ('S3Error',)
+
+from ..exceptions import ConveyorError
+
+
+class S3Error(ConveyorError):
+    """Error related to AWS S3 usage.
+    """
