+++ ltdconveyor/s3/utils.py
@@ -1,6 +1,8 @@
 """S3 utilities.
 """
 
+__all__ = ('open_bucket',)
+
 import boto3
